Track double to int

+++ ATM/Classes/Track.cs
@@ -9,13 +9,11 @@ namespace ATM.Classes
 {
     public class Track : ITrack
     {
-
         public int Tag { get; set; }
         public int Altitude { get; set; }
-        public double XCoordinate { get; set; }
-        public double YCoordinate { get; set; }
-        public double CurrentCompCourse { get; set; }
-        public double Velocity { get; set; }
-       
+        public int XCoordinate { get; set; }
+        public int YCoordinate { get; set; }
+        public int CurrentCompCourse { get; set; }
+        public int Velocity { get; set; }
     }
 }
